ENVI

Documentation for ENVI.

ENVI.envi_to_hdf5

ENVI.get_envi_params

ENVI.read_envi_file

ENVI.read_envi_header

ENVI.envi_to_hdf5 — Method

envi_to_hdf5(fpath::String, hdrpath::String, outpath::String)

Read ENVI formatted HSI file from fpath and its associated metadata file hdrpath. Save the array to an hdf5 file at outapth.

ENVI.get_envi_params — Method

get_envi_params(h::Dict)

Parse dict returned by read_envi_Header and return parameters needed for reading binary file.

ENVI.read_envi_file — Method

read_envi_file(fpath::String, hdrpath::String)

Read an ENVI formatted HSI file located at fpath with it's associated metatdata in hdrpath. Returns an image array img, the parsed header dictionary and the parameter dictionary.

ENVI.read_envi_header — Method

read_envi_header(file::String)

Reads and ENVI .hdr file header and returns the parameters in a dictionary as strings.